requirements.txt
# Core dependencies for transcriber web app
gradio>=4.0.0
openai>=1.0.0
pydub>=0.25.1
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Audio processing
ffmpeg-python>=0.2.0

# Type checking and validation
typing-extensions>=4.0.0

# JSON handling
json5>=0.9.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Code quality
ruff>=0.1.0
mypy>=1.0.0

# Additional utilities
requests>=2.28.0